Creation of OpenClosed principle
In this case the use of this principle is made to create the different classes to be able to keep adding functionality in them but without being able to modify the functionality already given to the code. In this case the filter is the chosen parameter to give functionality and apply this principle.