Creation of LiskovSubstitutionPrinciple
In this case the use of this principle is made to create the different classes from inheritance and being able of giving it values when some of the values are different from the parent like it's the case of the Square and the Rectangle. (In this case the Square gets both attributes values just by giving a value to 1 of them.